size is not an int
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@8185 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/3D/FontManager.cxx
@@ -400,7 +400,7 @@ float FontManager::getStrLength(int faceID, float size,	std::string text,
   if (!alreadyStripped)
     text = stripAnsiCodes(text);
 
-  return pFont->getStrLength(scale, text.c_str(), text.size());
+  return pFont->getStrLength(scale, text.c_str(), (int)text.size());
 }
 
 float FontManager::getStrLength(const std::string &face, float size,
